requirements.txt
amqp==5.1.1
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
celery==5.3.4
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==5.1
django-cors-headers==4.3.1
django-redis==5.4.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
kombu==5.3.2
msgpack==1.0.8
packaging==24.1
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.2
redis==5.0.6
requests==2.32.3
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.7.0
zope.interface==6.4.post2
